refactor: moves DropdownFilter to deprecated directory FE-2007
