+++ src/__deprecated__/components/dropdown-filter/dropdown-filter.scss
@@ -1,5 +1,5 @@
-@import "./../../style-config/colors";
-@import "./../../style-config/z-index";
+@import "./../../../style-config/colors";
+@import "./../../../style-config/z-index";
 
 .carbon-dropdown__list-item--no-results {
   font-style: italic;
