+++ src/__deprecated__/components/dropdown-filter/dropdown-filter.stories.js
@@ -8,13 +8,13 @@ import {
 } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import { State, Store } from '@sambego/storybook-state';
-import { classicThemeSelector } from '../../../.storybook/theme-selectors';
-import ImmutableHelper from '../../utils/helpers/immutable';
-import OptionsHelper from '../../utils/helpers/options-helper';
+import { classicThemeSelector } from '../../../../.storybook/theme-selectors';
+import ImmutableHelper from '../../../utils/helpers/immutable';
+import OptionsHelper from '../../../utils/helpers/options-helper';
 import notes from './documentation/notes.md';
 import Info from './documentation/Info';
 import DropdownFilter from './dropdown-filter';
-import getDocGenInfo from '../../utils/helpers/docgen-info';
+import getDocGenInfo from '../../../utils/helpers/docgen-info';
 
 DropdownFilter.__docgenInfo = getDocGenInfo(
   require('./docgenInfo.json'),
